# Conflicts:
#	crates/ruff/src/rules/ruff/rules/mod.rs
#	crates/ruff/src/rules/ruff/rules/mutable_defaults_in_class_fields.rs

+++ crates/ruff/src/rules/isort/mod.rs
@@ -52,7 +52,7 @@ pub enum AnnotatedImport<'a> {
     ImportFrom {
         module: Option<&'a str>,
         names: Vec<AnnotatedAliasData<'a>>,
-        level: Option<usize>,
+        level: Option<u32>,
         atop: Vec<Comment<'a>>,
         inline: Vec<Comment<'a>>,
         trailing_comma: TrailingComma,
