+++ crates/ruff_python_semantic/src/binding.rs
@@ -185,7 +185,7 @@ impl From<BindingId> for usize {
 #[derive(Debug, Clone)]
 pub struct StarImportation<'a> {
     /// The level of the import. `None` or `Some(0)` indicate an absolute import.
-    pub level: Option<usize>,
+    pub level: Option<u32>,
     /// The module being imported. `None` indicates a wildcard import.
     pub module: Option<&'a str>,
 }
